+++ x/observer/keeper/ballot.go
@@ -39,6 +39,10 @@ func (k Keeper) GetBallotList(ctx sdk.Context, height int64) (val types.BallotLi
 	return val, true
 }
 
+func (k Keeper) GetMaturedBallots(ctx sdk.Context, maturityBlocks int64) (val types.BallotListForHeight, found bool) {
+	return k.GetBallotList(ctx, ctx.BlockHeight()-maturityBlocks)
+}
+
 func (k Keeper) GetAllBallots(ctx sdk.Context) (voters []*types.Ballot) {
 	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.KeyPrefix(types.VoterKey))
 	iterator := sdk.KVStorePrefixIterator(store, []byte{})
@@ -59,14 +63,4 @@ func (k Keeper) AddBallotToList(ctx sdk.Context, ballot types.Ballot) {
 	}
 	list.BallotsIndexList = append(list.BallotsIndexList, ballot.BallotIdentifier)
 	k.SetBallotList(ctx, &list)
-}
-
-// GetMaturedBallotList Returns a list of ballots which are matured at current height
-func (k Keeper) GetMaturedBallotList(ctx sdk.Context) []string {
-	maturityBlocks := k.GetParamsIfExists(ctx).BallotMaturityBlocks
-	list, found := k.GetBallotList(ctx, ctx.BlockHeight()-maturityBlocks)
-	if !found {
-		return []string{}
-	}
-	return list.BallotsIndexList
 }